requirements.txt
alabaster==0.7.12
Babel==2.8.0
cacheout==0.11.2
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
config==0.5.0.post0
cryptography==3.0
decorator==4.4.2
docutils==0.16
idna==2.10
imagesize==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.4
pb-tool==3.1.0
progressbar2==3.51.4
pycparser==2.20
Pygments==2.6.1
pyotp==2.3.0
pyparsing==2.4.7
pyTelegramBotAPI==3.7.2
python-telegram-bot==12.8
python-utils==2.4.0
pytz==2020.1
requests==2.7.0
robin-stocks==1.4.2
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.2.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
tornado==6.0.4
urllib3==1.25.10
